+++ _IC50_Prediction/benchmark_test/TGSA_SANGER/preprocess_gene.py
@@ -57,6 +57,18 @@ def hugo_to_ncbi_map():
     return hugo_map
 
 
+def add_rand_noise(cell_data, std=1.0, noise_seed=2021) :
+    if noise_seed==-1 :
+        pass
+    else :
+        rng = np.random.default_rng(seed=noise_seed)
+        noise = rng.normal(0, std, size=cell_data.shape)
+        cell_data = cell_data + noise
+        print(f"# Adding noise N(0, {std}) to Cell Data...")
+        print(f"{noise}\n")
+    return cell_data
+
+
 # def save_cell_graph_exp(genes_path, save_path) :
 #     
 #     if not os.path.exists(save_path):
@@ -89,23 +101,57 @@ def save_cell_graph(genes_path, save_path) :
         os.makedirs(save_path)
     
     exp = pd.read_csv(os.path.join(genes_path, 'EXP.csv'), index_col=0)
-    cn = pd.read_csv(os.path.join(genes_path, 'CNV.csv'), index_col=0)
     mu = pd.read_csv(os.path.join(genes_path, 'MUT.csv'), index_col=0)
-
+    
+    cncat = "tcga" in genes_path
+    if not cncat:
+        # CNV in log2(CN-Ratio + 1)
+        cn = pd.read_csv(os.path.join(genes_path, 'CNV.csv'), index_col=0)
+    else:
+        # CNV categorized for prediction of TCGA dataset
+        cn = pd.read_csv(os.path.join(genes_path, 'CNV_Category.csv'), index_col=0)
+    
+    noise = "noise" in save_path
+    if noise : 
+        exp = add_rand_noise(exp)
+        mu = add_rand_noise(mu)
+        cn = add_rand_noise(cn)
+    
     index = exp.index
     columns = exp.columns
 
     scaler_exp = StandardScaler()
     scaler_cnv = StandardScaler()
+    
     exp = scaler_exp.fit_transform(exp)
     cn = scaler_cnv.fit_transform(cn)
+    
+    if genes_path == "./_data" and cncat:
+        file_scaler_exp = os.path.join(save_path, 'scaler_exp.pickle')
+        file_scaler_cnv = os.path.join(save_path, 'scaler_cncat.pickle')
+        
+        import pickle
+        with open(file_scaler_exp, "wb") as f:
+            pickle.dump(scaler_exp, f)
+        with open(file_scaler_cnv, "wb") as f:
+            pickle.dump(scaler_cnv, f)
+    
+    elif genes_path == "./_data" and not cncat:
+        file_scaler_cnv = os.path.join(save_path, 'scaler_cn.pickle')
+        
+        import pickle
+        with open(file_scaler_cnv, "wb") as f:
+            pickle.dump(scaler_cnv, f)
+    
+    else: pass
     
     imp_mean = SimpleImputer()
     exp = imp_mean.fit_transform(exp)
 
     exp = pd.DataFrame(exp, index=index, columns=columns)
     cn = pd.DataFrame(cn, index=index, columns=columns)
-    mu = pd.DataFrame(mu, index=index, columns=columns)
+    # mu = pd.DataFrame(mu, index=index, columns=columns)
+    mu = pd.DataFrame(mu.values, index=index, columns=columns)
     
     cell_dict = {}
     cell_names = exp.index
@@ -113,7 +159,10 @@ def save_cell_graph(genes_path, save_path) :
     for i in cell_names:
         cell_dict[i] = Data(x=torch.tensor([exp.loc[i], cn.loc[i], mu.loc[i]], dtype=torch.float).T)
     
-    np.save(os.path.join(save_path, 'cell_feature_all.npy'), cell_dict)
+    if not cncat:
+        np.save(os.path.join(save_path, 'cell_feature_all.npy'), cell_dict)
+    else:
+        np.save(os.path.join(save_path, 'cell_feature_all_cncat.npy'), cell_dict)
 
 
 def get_STRING_graph(genes_path, thresh=0.95):
@@ -183,7 +232,10 @@ if __name__ == '__main__':
     
     # genes_path = "./_data"
     # genes_path = "./_data_gdsc"
-    genes_path = "./_data_ccle"
+    # genes_path = "./_data_ccle"
+    # genes_path = "./_data_tcga"
+    # genes_path = "./_data_liu24_invivo"
+    genes_path = "./_data_noise"
     save_path = genes_path
     
     # edge_index_850 = get_STRING_graph(genes_path, thresh=0.85)
@@ -198,16 +250,26 @@ if __name__ == '__main__':
     
     
     # EXP for Similarity Graph
-    import pickle
-    file1 = os.path.join(genes_path, 'EXP.csv')
-    file2 = os.path.join(genes_path, 'EXP_Total_Scaled.csv')
-    
-    exp = pd.read_csv(file1, index_col=0)
-    exp_scaled = pd.read_csv(file2, index_col=0)
-    
-    sum(exp_scaled.index!=exp.index)   # 0
-    sum(exp_scaled.index==exp.index)   # 1399
-    
-    cell_feature_normalized = {k:v for k,v in zip(exp_scaled.index, exp_scaled.values)}
-    with open("./{}/cell_feature_normalized".format(save_path), "wb") as f : 
-        pickle.dump(cell_feature_normalized, f)
+    if genes_path not in ["./_data_tcga", "./_data_liu24_invivo"]:
+        file = "./{}/cell_feature_normalized".format(save_path)
+        if os.path.exists(file):
+            print("Normalized Feature Exist...")
+        else:
+            import pickle
+            file1 = os.path.join(genes_path, 'EXP.csv')
+            file2 = os.path.join(genes_path, 'EXP_Total_Scaled.csv')
+            
+            exp = pd.read_csv(file1, index_col=0)
+            exp_scaled = pd.read_csv(file2, index_col=0)
+            
+            noise = "noise" in save_path
+            if noise : 
+                exp = add_rand_noise(exp)
+                exp_scaled = add_rand_noise(exp_scaled)
+            
+            sum(exp_scaled.index!=exp.index)   # 0
+            sum(exp_scaled.index==exp.index)   # 1399
+            
+            cell_feature_normalized = {k:v for k,v in zip(exp_scaled.index, exp_scaled.values)}
+            with open(file, "wb") as f : 
+                pickle.dump(cell_feature_normalized, f)
